Scripts/YouTube/requirements.txt
google_api_python_client==2.40.0
google_auth_oauthlib==0.5.0
nltk==3.7
numpy==1.22.2
pandas==1.4.1
protobuf==3.19.4
scikit_learn==1.0.2
youtube_transcript_api==0.4.3
